Add rest of patient functionality
User logs in (currently with both ssn and name to be changed).
If the patient exists then create session keys and go to patientlanding.
If the patient does not exist, go to register then patientlanding.
At patientlanding, ask for treatment wanted, then redirect to filter
for doctors that give that treatment, then redirect to filter for
hospitals that the doctor works at, then redirect to filter for
schedules matching all criteria. Upon selection of appointment,
insert appointment into table.

TODO:
- Display all patient appointments on patientlanding.
- Edit login for just ssn and query for patient name using ssn at
  patientlanding.